+++ repository-jpa/src/test/java/ru/dpopkov/knownics/repository/jpa/DataModuleTestConfiguration.java
@@ -1,0 +1,9 @@
+package ru.dpopkov.knownics.repository.jpa;
+
+import org.springframework.boot.SpringBootConfiguration;
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+
+@SpringBootConfiguration
+@EnableAutoConfiguration
+public class DataModuleTestConfiguration {
+}