+++ history.md
@@ -5,6 +5,7 @@ History of implementation
 * Add module repository-jpa. Closes #11.
 * Add Profiles. Closes #12.
 * Add data.sql to populate database in repository-jpa profile.
+* Add findByDescription to CategoryRepository, add Integration Test for CategoryJpaRepository.
 ##### 2021-12-26
 * Add module services. Add CategoryService, CategoryServiceImpl.
 * Implement equals and hashCode for entities. Closes #9.